+++ core/src/main/java/edu/berkeley/cs/succinct/buffers/SuccinctBuffer.java
@@ -233,7 +233,7 @@ public class SuccinctBuffer extends SuccinctCore {
    * @param i Index into inverted alphabet map
    * @return Value of inverted alphabet map at specified index.
    */
-  @Override public long lookupC(long i) {
+  @Override public int lookupC(long i) {
     return SerializedOperations.ArrayOps.getRank1(coloffsets.buffer(), 0, getSigmaSize(), i) - 1;
   }
